+++ lib/Importer.js
@@ -1,32 +1,13 @@
 'use strict';
 
-const StringScanner = require('StringScanner');
 const eslint = require('eslint');
-const lodashRange = require('lodash.range');
 
 const CommandLineEditor = require('./CommandLineEditor');
 const Configuration = require('./Configuration');
-const ImportStatement = require('./ImportStatement');
 const ImportStatements = require('./ImportStatements');
 const JsModule = require('./JsModule');
+const findCurrentImports = require('./findCurrentImports');
 const findJsModulesFor = require('./findJsModulesFor');
-const xRegExp = require('./xregexp');
-
-const REGEX_SKIP_SECTION = xRegExp(`
-  ^
-  \\s*                     # preceding whitespace
-  (?:
-    (?<quote>['"])use\\sstrict\\k<quote>;? # 'use strict';
-    |
-    //[^\n]*               # single-line comment
-    |
-    /\\*                   # open multi-line comment
-    (?:\n|.)*?             # inside of multi-line comment
-    \\*/                   # close multi-line comment
-  )?                       # ? b/c we want to match on only whitespace
-  \n                       # end of line
-  `, 'xs' // free-spacing, dot-match-all
-);
 
 class Importer {
   constructor(lines, pathToCurrentFile) {
@@ -323,43 +304,7 @@ class Importer {
    * @return {Object}
    */
   findCurrentImports() {
-    /* eslint-disable no-cond-assign */
-    let importsStartAtLineNumber = 1;
-    let newlineCount = 0;
-
-    const scanner = new StringScanner(this.editor.currentFileContent());
-    let skipped = '';
-    let skipSection;
-    while (skipSection = scanner.scan(REGEX_SKIP_SECTION)) {
-      skipped += skipSection;
-    }
-
-    // We don't want to skip over blocks that are only whitespace
-    if (xRegExp.test(skipped, xRegExp('^\\s+$', 's'))) {
-      scanner.reset();
-    } else {
-      const count = (skipped.match(/\n/g) || []).length;
-      importsStartAtLineNumber = importsStartAtLineNumber + count;
-    }
-
-    const imports = new ImportStatements(this.config);
-    let potentialImport;
-    while (potentialImport = scanner.scan(xRegExp('(^\\s*\\n)*^.*?;\\n', 's'))) {
-      const importStatement = ImportStatement.parse(potentialImport.trim());
-      if (!importStatement) {
-        break;
-      }
-
-      imports.push(importStatement);
-      const count = (potentialImport.match(/\n/g) || []).length;
-      newlineCount += count;
-    }
-
-    const importsEndAtLineNumber = importsStartAtLineNumber + newlineCount + 1;
-    return {
-      imports,
-      range: lodashRange(importsStartAtLineNumber, importsEndAtLineNumber),
-    };
+    return findCurrentImports(this.config, this.editor.currentFileContent());
   }
 
   /**
